+++ app/services/NotificationService.scala
@@ -7,6 +7,7 @@ import models._
 import play.api.Logger
 import play.api.libs.mailer.MailerClient
 import play.api.libs.mailer.Email
+import play.api.mvc.{AnyContent, Request}
 
 @Singleton
 class NotificationService @Inject()(configuration: play.api.Configuration,
@@ -47,20 +48,41 @@ class NotificationService @Inject()(configuration: play.api.Configuration,
     }
   }
 
+  def newLoginRequest(request: Request[AnyContent], user: User) = {
+    val url = s"${routes.HomeController.index().absoluteURL()(request)}?key=${user.key}"
+    val bodyHtml = s"""Bonjour ${user.name},<br>
+                      |<br>
+                      |Vous pouvez vous connecter au service A+ en ouvrant l'adresse suivante :<br>
+                      |<a href="${url}">${url}</a>
+                      |<br>
+                      |<br>
+                      |<b>Ce mail est personnel, ne le transférez pas, il permettrait à quelqu'un d'autre d'utiliser votre identité sur le réseau A+.</b>
+                      |<br>
+                      |<br>
+                      |Merci de votre aide,<br>
+                      |Si vous avez des questions, n'hésitez pas à nous contacter sur contact@aplus.beta.gouv.fr<br>
+                      |Equipe A+""".stripMargin
+    val email = play.api.libs.mailer.Email(
+      s"Connexion à A+",
+      from = "A+ Notification: ne pas répondre <ne-pas-repondre@aplus.beta.gouv.fr>",
+      Seq(s"${user.name} <${user.email}>"),
+      bodyHtml = Some(bodyHtml)
+    )
+    sendMail(email)
+  }
+
+
   private def generateFooter(user: User): String = {
     val delegates = if(user.delegations.nonEmpty) {
-      s"<br><br>Les personnes suivantes ont une délégations sur votre compte agent : <b>${user.delegations.map { case (name, email) => s"$name <$email>" }.mkString(", ")}</b>. Elles reçoivent une copie de ce mail et peuvent agir en votre nom sur le réseau A+.<br>"
+      s"- Les personnes suivantes ont une délégations sur votre compte agent : <b>${user.delegations.map { case (name, email) => s"$name &#x3C;$email&#x3E;" }.mkString(", ")}</b>. (Elles peuvent agir en votre nom sur le réseau A+)<br>"
     } else {
       ""
     }
     return s"""<br>
-       |<b>Ne transférez pas cet email, il vous est personnellement destiné, il permettrait à quelqu'un d'autre d'utiliser votre identité sur le réseau A+.</b>
-       | Si vous souhaitez transférer la demande à quelqu'un d'autre vous avez la possibilité de le faire en ouvrant le lien ci-dessus.
-       | $delegates
-       |<br>
-       |<br>
-       |<b>Ne pas répondre à ce mail.</b><br>
-       |<i>Si vous avez des questions à propos de l'outil A+, contactez-nous sur contact@aplus.beta.gouv.fr</i><br>
+       |<b>Ne transférez pas cet email et ne pas répondre à cet email</b><br><i>
+       |$delegates
+       |- Vous pouvez transférer la demande à un autre agent en ouvrant le lien ci-dessus<br>
+       |- Si vous avez un problème ou besoins d'aide à propos de l'outil A+, contactez-nous sur contact@aplus.beta.gouv.fr</i>
      """.stripMargin
   }
 
@@ -70,16 +92,14 @@ class NotificationService @Inject()(configuration: play.api.Configuration,
     val bodyHtml = s"""Bonjour ${invitedUser.name},<br>
                       |<br>
                       |<p>${answer.map(_.creatorUserName).getOrElse(application.creatorUserName)} a besoin de vous.<br>
-                      |Cette personne vous a invité sur la demande suivante: "${application.subject}"<br>
-                      |<q>${application.description}</q><p><br>
-                      |<br>
-                      |Vous pouvez accéder à la demande en suivant ce lien: <br>
+                      |Cette personne vous a invité sur la demande suivante: "${application.subject}"
+                      |<br><q><i>${application.description}</i></q></p>
+                      |Vous pouvez voir la demande et y répondre en suivant ce lien: <br>
                       |<a href="${url}">${url}</a><br>
-                      |<br>
                       |$footer
                       """.stripMargin
     Email(subject = s"[A+] Nouvelle demande d'aide : ${application.subject}",
-      from = "A+ <contact@aplus.beta.gouv.fr>",
+      from = "A+ Notification: ne pas répondre <ne-pas-repondre-aplus@beta.gouv.fr>",
       to = List(s"${invitedUser.name} <${invitedUser.email}>"),
       cc = invitedUser.delegations.map { case (name, email) => s"$name <$email>" }.toSeq,
       bodyHtml = Some(bodyHtml)
@@ -91,16 +111,14 @@ class NotificationService @Inject()(configuration: play.api.Configuration,
     val footer = generateFooter(user)
     val bodyHtml = s"""Bonjour ${user.name},<br>
                       |<br>
-                      |<p>${answer.creatorUserName} a donné une réponse sur la demande: "${application.subject}"<br>
-                      |<q>${answer.message}</q></p>
-                      |<br>
-                      |Vous pouvez accéder à l'ensemble de la demande en suivant ce lien: <br>
+                      |<p>${answer.creatorUserName} a donné une réponse sur la demande: "${application.subject}"
+                      |<br><q><i>${answer.message}</q></p>
+                      |Vous pouvez voir la demande et y répondre en suivant ce lien: <br>
                       |<a href="${url}">${url}</a>
-                      |<br>
                       |$footer
                       """.stripMargin
     Email(subject = s"[A+] Nouvelle réponse pour: ${application.subject}",
-      from = "A+ <contact@aplus.beta.gouv.fr>",
+      from = "A+ Notification: ne pas répondre <ne-pas-repondre-aplus@beta.gouv.fr>",
       to = List(s"${user.name} <${user.email}>"),
       cc = user.delegations.map { case (name, email) => s"$name <$email>" }.toSeq,
       bodyHtml = Some(bodyHtml)
